Clear tile flags from the gid before further processing
Better to show non-rotated/flipped tiles than no tile at all. This also
fixes interpretation of collision tiles that happen to be flipped.

Also interpret the gid as an unsigned number, since that's how they are
written in the TMX file since the introduction of these flags.

+++ src/utils/xml.cpp
@@ -99,7 +99,7 @@ namespace XML
         xmlChar *prop = xmlGetProp(node, BAD_CAST name);
         if (prop)
         {
-            ret = atoi((char*)prop);
+            ret = atol((char*)prop);
             xmlFree(prop);
         }
